requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
openai>=1.0.0
python-dotenv>=1.0.0
langsmith>=0.2.0
numpy>=1.24.0
pypdf>=4.0.0
rank-bm25>=0.2.2
